get changes from adafruit

+++ actions_install.sh
@@ -4,11 +4,15 @@ set -e
 
 pip3 install clint pyserial setuptools adafruit-nrfutil
 sudo apt-get update
+sudo apt-get install -y --allow-downgrades libllvm8=1:8-3~ubuntu18.04.2 -V
 sudo apt install -fy cppcheck clang-format-8
-sudo ln -s /usr/bin/clang-format-8 /usr/bin/clang-format
+if [ ! -f /usr/bin/clang-format ]; then
+    sudo ln -s /usr/bin/clang-format-8 /usr/bin/clang-format
+fi
 
 # make all our directories we need for files and libraries
 mkdir ${HOME}/.arduino15
 mkdir ${HOME}/.arduino15/packages
 mkdir ${HOME}/Arduino
 mkdir ${HOME}/Arduino/libraries
+
